requirements.txt
annotated-types==0.6.0
ansible==8.5.0
ansible-core==2.15.5
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.1.1
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
distro==1.8.0
dotty-dict==1.3.1
GDAL==3.6.4
greenlet==2.0.2
h11==0.14.0
halo==0.0.31
hid==1.0.5
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
log-symbols==0.0.14
MarkupSafe==2.1.3
meson==1.2.0
milc==1.6.8
msgpack==1.0.4
numpy==1.25.1
openai==1.2.3
outcome==1.3.0.post0
packaging==23.2
paramiko==3.3.1
Pillow==10.0.0
protobuf==4.23.4
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.15.1
PyNaCl==1.5.0
pynvim==0.4.3
pyserial==3.5
PySocks==1.7.1
python-dotenv==1.0.1
pyusb==1.2.1
PyYAML==6.0.1
qmk==1.1.2
rapidfuzz==3.6.1
referencing==0.29.1
requests==2.31.0
resolvelib==1.0.1
rpds-py==0.8.11
selenium==4.18.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spinners==0.0.24
termcolor==2.3.0
thefuzz==0.22.1
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
